move datachecks to torch/ and fix up docs + formatting
