+++ src/poprox_recommender/torch/__init__.py
@@ -1,3 +1,7 @@
 """
 Utility functions for implementing POPROX recommender components with PyTorch.
 """
+
+from .datachecks import assert_tensor_size
+
+__all__ = ["assert_tensor_size"]
